# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vespera luxury minaudières e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CollectionPiece {
  id: number;
  title: string;
  description: string;
  price: number;
  stockCount: number;
  primaryImage: string;
  images: string[];
  material: string;
  dimensions: string;
  occasionStyling: string[];
  artisanNotes: string;
  isFeatured: boolean;
  slug: string;
}

export interface AdminProductInput {
  title: string;
  description: string;
  price: number;
  stockCount?: number;
  primaryImage: string;
  images?: string[];
  material?: string;
  dimensions?: string;
  occasionStyling?: string[];
  artisanNotes?: string;
  isFeatured?: boolean;
  slug: string;
}

export interface CartItem {
  pieceId: number;
  quantity: number;
}

export interface CheckoutRequest {
  items: CartItem[];
}

export interface CheckoutResponse {
  orderId: string;
  keyId: string;
  amount: number;
  currency: string;
}

export interface VerifyPaymentRequest {
  razorpay_order_id: string;
  razorpay_payment_id: string;
  razorpay_signature: string;
}

export interface VerifyPaymentResponse {
  verified: boolean;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorResponse {
  error: string;
}

export type AdminDeleteProduct200 = {
  success: boolean;
};