# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vespera luxury minaudières e-commerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: collection
    description: Product collection operations
  - name: cart
    description: Cart and checkout operations
  - name: admin
    description: Admin management operations
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /collection:
    get:
      operationId: listCollection
      tags: [collection]
      summary: List all pieces in the collection
      description: Returns all available minaudières in the collection
      responses:
        "200":
          description: Collection pieces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionPiece"

  /collection/{id}:
    get:
      operationId: getCollectionPiece
      tags: [collection]
      summary: Get a single collection piece
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection piece
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionPiece"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /collection/featured:
    get:
      operationId: getFeaturedPieces
      tags: [collection]
      summary: Get featured pieces for the home page
      description: Returns a curated selection of featured pieces for the entrance
      responses:
        "200":
          description: Featured pieces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionPiece"

  /checkout:
    post:
      operationId: createCheckoutSession
      tags: [cart]
      summary: Create a Razorpay checkout order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Razorpay order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /checkout/verify:
    post:
      operationId: verifyPayment
      tags: [cart]
      summary: Verify Razorpay payment signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Payment verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPaymentResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/collection:
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionPiece"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/collection/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update an existing product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionPiece"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset from PUBLIC_OBJECT_SEARCH_PATHS
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CollectionPiece:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        stockCount:
          type: integer
        primaryImage:
          type: string
        images:
          type: array
          items:
            type: string
        material:
          type: string
        dimensions:
          type: string
        occasionStyling:
          type: array
          items:
            type: string
        artisanNotes:
          type: string
        isFeatured:
          type: boolean
        slug:
          type: string
      required:
        - id
        - title
        - description
        - price
        - stockCount
        - primaryImage
        - images
        - material
        - dimensions
        - occasionStyling
        - artisanNotes
        - isFeatured
        - slug

    AdminProductInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        stockCount:
          type: integer
        primaryImage:
          type: string
        images:
          type: array
          items:
            type: string
        material:
          type: string
        dimensions:
          type: string
        occasionStyling:
          type: array
          items:
            type: string
        artisanNotes:
          type: string
        isFeatured:
          type: boolean
        slug:
          type: string
      required:
        - title
        - description
        - price
        - primaryImage
        - slug

    CheckoutRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
      required:
        - items

    CartItem:
      type: object
      properties:
        pieceId:
          type: integer
        quantity:
          type: integer
      required:
        - pieceId
        - quantity

    CheckoutResponse:
      type: object
      properties:
        orderId:
          type: string
        keyId:
          type: string
        amount:
          type: integer
        currency:
          type: string
      required:
        - orderId
        - keyId
        - amount
        - currency

    VerifyPaymentRequest:
      type: object
      properties:
        razorpay_order_id:
          type: string
        razorpay_payment_id:
          type: string
        razorpay_signature:
          type: string
      required:
        - razorpay_order_id
        - razorpay_payment_id
        - razorpay_signature

    VerifyPaymentResponse:
      type: object
      properties:
        verified:
          type: boolean
      required:
        - verified

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error