# lib/api-zod/src/generated/types/adminProductInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vespera luxury minaudières e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminProductInput {
  title: string;
  description: string;
  price: number;
  stockCount?: number;
  primaryImage: string;
  images?: string[];
  material?: string;
  dimensions?: string;
  occasionStyling?: string[];
  artisanNotes?: string;
  isFeatured?: boolean;
  slug: string;
}