# lib/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vespera luxury minaudières e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlRequest } from "./uploadUrlRequest";

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}